# [internal] Improve performance of Go dependency inference