Closes https://github.com/pantsbuild/pants/issues/12928#issuecomment-934805922.

This also changes our behavior to stop caring about `TestImports` for `_go_external_package` targets, given that we'll never test those.

### `dependencies` benchmark

This requires changing this line to use `Targets` so that generated targets have their dependencies resolved too:

https://github.com/pantsbuild/pants/blob/a744978d2badd06509cc6e8091c14220bed3aff6/src/python/pants/backend/project_info/dependencies.py#L94

Before (I aborted after 1 run because it was so slow): 

```
❯ hyperfine -r 5 './pants_from_sources --no-pantsd --no-process-execution-local-cache dependencies ::'
Current estimate: 344.640 s
```

After:

```
❯ /usr/bin/time ./pants_from_sources --no-pantsd --no-process-execution-local-cache dependencies ::
       27.25 real        30.56 user        24.23 sys
```

Almost all of the time was coming from here, with up to 10 seconds per package. I don't know why it's so pathologically slow - we set `GOPROXY=off` so we should be confident that nothing is being downloaded.

https://github.com/pantsbuild/pants/blob/a744978d2badd06509cc6e8091c14220bed3aff6/src/python/pants/backend/go/util_rules/external_module.py#L206-L240 

### `package` benchmark

Before:

```
❯ hyperfine -r 5 './pants_from_sources --no-pantsd --no-process-execution-local-cache package ::'
Benchmark #1: ./pants_from_sources --no-pantsd --no-process-execution-local-cache package ::
  Time (mean ± σ):     54.981 s ±  1.441 s    [User: 70.577 s, System: 81.823 s]
  Range (min … max):   53.426 s … 57.188 s    5 runs
```

After:

```
❯ hyperfine -r 5 './pants_from_sources --no-pantsd --no-process-execution-local-cache package ::'
  Time (mean ± σ):     50.658 s ±  0.700 s    [User: 66.903 s, System: 74.338 s]
  Range (min … max):   49.611 s … 51.245 s    5 runs
```


[ci skip-rust]
[ci skip-build-wheels]