@Eric-Arellano 2021-10-05
> An alternative to storing the metadata on the target itself is to rely on Pantsd memoization. That might be cleaner?

I'm gonna try this approach in a different PR. Because target generation isn't saved on disk, there's no benefit fwict to storing it on the target. I don't think there's value to `./pants peek` letting you see these implementation details either.