@Eric-Arellano 2021-10-05
> I'm gonna try this approach in a different PR.

This is a much better factoring. Going to close this in favor of that.